fix: use new dropdown component
